requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
fonttools==4.44.0
frozenlist==1.4.0
idna==3.4
kiwisolver==1.4.5
magic-filter==1.0.12
matplotlib==3.8.1
multidict==6.0.4
numpy==1.26.1
packaging==23.2
Pillow==10.1.0
pydantic==2.3.0
pydantic_core==2.6.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.7
yarl==1.9.2
